Ported column alias tests
